+++ Code_parachute/mainpara.py
@@ -49,8 +49,8 @@ class Mainpara(Parachute):
 
         """This computation is based on Knacke's book. It is valid for a conical ribbon parachute"""
         rho = coesa76(z / 1000).rho
-        self.opening_force = self.cd * self.surface * (
-                    1 / 2 * rho * v**2) * self.cx_factor * self.x1_factor
+        self.opening_force = float(self.cd * self.surface * (
+                    1 / 2 * rho * v**2) * self.cx_factor * self.x1_factor)
 
         if self.opening_force > 1e5:
             print("Warning: your main parachute would probably be in pieces!\n")
